change: Move the getBoard function to LGFX_Device.
