+++ src/M5UnitLCD.h
@@ -6,8 +6,9 @@
 // #include <SD.h>
 // #include <SPIFFS.h>
 // #include <HTTPClient.h>
-#include "M5GFX.h"
+
 #include "lgfx/v1/panel/Panel_M5UnitLCD.hpp"
+#include "M5GFX.h"
 
 #if defined ( ARDUINO )
  #include <Arduino.h>
@@ -49,8 +50,6 @@ class M5UnitLCD : public lgfx::LGFX_Device
   lgfx::Panel_M5UnitLCD _panel_instance;
 
 public:
-
-  m5gfx::board_t getBoard(void) const { return m5gfx::board_t::board_M5UnitLCD; }
 
   M5UnitLCD(uint8_t pin_sda = M5UNITLCD_SDA, uint8_t pin_scl = M5UNITLCD_SCL, uint32_t i2c_freq = M5UNITLCD_FREQ, int8_t i2c_port = -1, uint8_t i2c_addr = M5UNITLCD_ADDR)
   {
@@ -102,6 +101,7 @@ public:
     }
 
     setPanel(&_panel_instance);
+    _board = lgfx::board_t::board_M5UnitLCD;
   }
 };
